feat: add role owner + update delete tables & menus

+++ app/Http/Controllers/api/v1/MenusController.php
@@ -108,11 +108,25 @@ class MenusController extends Controller
         }
     }
 
-    public function destroy(Menus $menu)
+    public function destroy($id)
     {
         try {
+            $menu = Menus::find($id);
+
+            if (!$menu) {
+                return response()->json([
+                    'message' => 'Menu not found, nothing to delete.',
+                ], 204);
+            }
+
+            if ($menu->image && str_contains($menu->image, 'amazonaws.com')) {
+                $oldPath = parse_url($menu->image, PHP_URL_PATH);
+                $oldKey = ltrim($oldPath, '/');
+                Storage::disk('s3')->delete($oldKey);
+            }
+
             $menu->delete();
-    
+
             return response()->json([
                 'message' => 'Menu deleted successfully.',
             ], 204);
